Update the restoring the language in dashboard
