+++ src/app/ng5-basic/nav-bar/nav-bar.component.html
@@ -3,7 +3,7 @@
       <div class="active-application-wrapper">
         <div class="active-application" title="Work Office In London">
           <div class="application-avatar">WO</div>
-          <div class="application-title">{{CurrentLanguage()}}</div>
+          <div class="application-title">{{CurrentLanguage()}} - <a class="change-language" routerLink="/profile">Change</a></div>
         </div>
       </div>
       <a class="navbar-brand">IOT Angular Dashboard</a>
