+++ src/app/ng5-basic/nav-bar/nav-bar.component.scss
@@ -87,3 +87,6 @@
         font-size: 1.25rem;
     }
 }
+.change-language { 
+  font-size: 12px;
+}
