+++ src/app/ng5-basic/nav-bar/nav-bar.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { SidebarControllerService } from '../ngx-sidebar/sidebar-controller.service';
 import { TranslateService } from '@ngx-translate/core';
+import { SupportedLanguages } from '@app/services/globalization.service';
 
 @Component({
   selector: 'app-nav-bar',
@@ -21,6 +22,6 @@ export class NavBarComponent implements OnInit {
     this.sidebar.ToggleSidebar.emit();
   }
   public CurrentLanguage () {
-    return this.translate.currentLang;
+    return SupportedLanguages[this.translate.currentLang];
   }
 }
